+++ social_core/backends/base.py
@@ -13,6 +13,7 @@ class BaseAuth(object):
     supports_inactive_user = False  # Django auth
     ID_KEY = None
     EXTRA_DATA = None
+    GET_ALL_EXTRA_DATA = False
     REQUIRES_EMAIL_VALIDATION = False
     SEND_USER_AGENT = False
     SSL_PROTOCOL = None
@@ -116,7 +117,12 @@ class BaseAuth(object):
             # store the last time authentication toke place
             'auth_time': int(time.time())
         }
-        for entry in (self.EXTRA_DATA or []) + self.setting('EXTRA_DATA', []):
+        extra_data_entries = []
+        if self.GET_ALL_EXTRA_DATA or self.setting('GET_ALL_EXTRA_DATA', False):
+            extra_data_entries = response.keys()
+        else:
+            extra_data_entries = (self.EXTRA_DATA or []) + self.setting('EXTRA_DATA', [])
+        for entry in extra_data_entries:
             if not isinstance(entry, (list, tuple)):
                 entry = (entry,)
             size = len(entry)