+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added Udata OAuth2 backend
 - Added ORCID backend
+- Added feature to get all extra data from backend through `GET_ALL_EXTRA_DATA` boolean flag.
 
 ## [1.3.0](https://github.com/python-social-auth/social-core/releases/tag/1.3.0) - 2017-05-06